NexOS v1.01.00 Release
This version now has support for an IO_BUFFER in general.  The PIC32 port supports the UART, CAN, and ADC modules for use with an IO_BUFFER.  There were more EVENTs added along with additional callbacks.  Also, a large amount of bug fixes.  From this release on, much more strict release notes will be made, and a logging system for the bugs.